+++ cmd/aws-sso/main.go
@@ -28,6 +28,7 @@ import (
 	// "github.com/davecgh/go-spew/spew"
 	"github.com/sirupsen/logrus"
 	"github.com/synfinatic/aws-sso-cli/internal/awscreds"
+	"github.com/synfinatic/aws-sso-cli/internal/ecs"
 	"github.com/synfinatic/aws-sso-cli/internal/ecs/client"
 	"github.com/synfinatic/aws-sso-cli/internal/ecs/server"
 	"github.com/synfinatic/aws-sso-cli/internal/helper"
@@ -148,6 +149,7 @@ func main() {
 	tags.SetLogger(log)
 	url.SetLogger(log)
 	utils.SetLogger(log)
+	ecs.SetLogger(log)
 	server.SetLogger(log)
 	client.SetLogger(log)
